Reorganize functions into more logical endpoints
