+++ src/routes/api/v1/metadata-map/+server.ts
@@ -1,0 +1,9 @@
+import { json } from "@sveltejs/kit";
+import type { RequestHandler } from "@sveltejs/kit";
+import { loadPageMetadataMap } from "$lib/loadPageMetadataMap";
+
+export const GET = (async () => {
+  const { pageMetadataMap } = await loadPageMetadataMap({ includeBreadcrumbs: false });
+  // exclude the home page and pages where we can't determine the URL
+  return json(Array.from(pageMetadataMap.values()).filter((page) => page.url && !page.isRoot));
+}) satisfies RequestHandler;
