+++ src/routes/api/v1/search-index/reset/+server.ts
